+++ pallets/subtensor/src/staking/add_stake.rs
@@ -1,0 +1,125 @@
+use super::*;
+use frame_support::{
+    storage::IterableStorageDoubleMap,
+    traits::{
+        tokens::{
+            fungible::{Balanced as _, Inspect as _, Mutate as _},
+            Fortitude, Precision, Preservation,
+        },
+        Imbalance,
+    },
+};
+
+impl<T: Config> Pallet<T> {
+    /// ---- The implementation for the extrinsic add_stake: Adds stake to a hotkey account.
+    ///
+    /// # Args:
+    /// * 'origin': (<T as frame_system::Config>RuntimeOrigin):
+    ///     -  The signature of the caller's coldkey.
+    ///
+    /// * 'hotkey' (T::AccountId):
+    ///     -  The associated hotkey account.
+    ///
+    /// * 'stake_to_be_added' (u64):
+    ///     -  The amount of stake to be added to the hotkey staking account.
+    ///
+    /// # Event:
+    /// * StakeAdded;
+    ///     -  On the successfully adding stake to a global account.
+    ///
+    /// # Raises:
+    /// * 'NotEnoughBalanceToStake':
+    ///     -  Not enough balance on the coldkey to add onto the global account.
+    ///
+    /// * 'NonAssociatedColdKey':
+    ///     -  The calling coldkey is not associated with this hotkey.
+    ///
+    /// * 'BalanceWithdrawalError':
+    ///     -  Errors stemming from transaction pallet.
+    ///
+    /// * 'TxRateLimitExceeded':
+    ///     -  Thrown if key has hit transaction rate limit
+    ///
+    pub fn do_add_stake(
+        origin: T::RuntimeOrigin,
+        hotkey: T::AccountId,
+        stake_to_be_added: u64,
+    ) -> dispatch::DispatchResult {
+        // We check that the transaction is signed by the caller and retrieve the T::AccountId coldkey information.
+        let coldkey = ensure_signed(origin)?;
+        log::info!(
+            "do_add_stake( origin:{:?} hotkey:{:?}, stake_to_be_added:{:?} )",
+            coldkey,
+            hotkey,
+            stake_to_be_added
+        );
+
+        // Ensure the callers coldkey has enough stake to perform the transaction.
+        ensure!(
+            Self::can_remove_balance_from_coldkey_account(&coldkey, stake_to_be_added),
+            Error::<T>::NotEnoughBalanceToStake
+        );
+
+        // Ensure that the hotkey account exists this is only possible through registration.
+        ensure!(
+            Self::hotkey_account_exists(&hotkey),
+            Error::<T>::HotKeyAccountNotExists
+        );
+
+        // Ensure that the hotkey allows delegation or that the hotkey is owned by the calling coldkey.
+        ensure!(
+            Self::hotkey_is_delegate(&hotkey) || Self::coldkey_owns_hotkey(&coldkey, &hotkey),
+            Error::<T>::HotKeyNotDelegateAndSignerNotOwnHotKey
+        );
+
+        // Ensure we don't exceed stake rate limit
+        let stakes_this_interval =
+            Self::get_stakes_this_interval_for_coldkey_hotkey(&coldkey, &hotkey);
+        ensure!(
+            stakes_this_interval < Self::get_target_stakes_per_interval(),
+            Error::<T>::StakeRateLimitExceeded
+        );
+
+        // If this is a nomination stake, check if total stake after adding will be above
+        // the minimum required stake.
+
+        // If coldkey is not owner of the hotkey, it's a nomination stake.
+        if !Self::coldkey_owns_hotkey(&coldkey, &hotkey) {
+            let total_stake_after_add =
+                Stake::<T>::get(&hotkey, &coldkey).saturating_add(stake_to_be_added);
+
+            ensure!(
+                total_stake_after_add >= NominatorMinRequiredStake::<T>::get(),
+                Error::<T>::NomStakeBelowMinimumThreshold
+            );
+        }
+
+        // Ensure the remove operation from the coldkey is a success.
+        let actual_amount_to_stake =
+            Self::remove_balance_from_coldkey_account(&coldkey, stake_to_be_added)?;
+
+        // If we reach here, add the balance to the hotkey.
+        Self::increase_stake_on_coldkey_hotkey_account(&coldkey, &hotkey, actual_amount_to_stake);
+
+        // Set last block for rate limiting
+        let block: u64 = Self::get_current_block_as_u64();
+        Self::set_last_tx_block(&coldkey, block);
+
+        // Emit the staking event.
+        Self::set_stakes_this_interval_for_coldkey_hotkey(
+            &coldkey,
+            &hotkey,
+            stakes_this_interval.saturating_add(1),
+            block,
+        );
+        log::info!(
+            "StakeAdded( hotkey:{:?}, stake_to_be_added:{:?} )",
+            hotkey,
+            actual_amount_to_stake
+        );
+        Self::deposit_event(Event::StakeAdded(hotkey, actual_amount_to_stake));
+
+        // Ok and return.
+        Ok(())
+    }
+}
