+++ pallets/subtensor/src/staking/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+pub mod helpers;
+pub mod add_stake;
+pub mod remove_stake;
+pub mod decrease_take;
+pub mod increase_take;
+pub mod become_delegate;
